debug class created and added, beginning of adding debug messages to other classes and beginning of automatically logging all errors to text file

+++ New Project/TD2DE_Texture.h
@@ -50,7 +50,7 @@ namespace TD2DE
 	public:
 		int GetTextureLocation(std::string textureName);
 		TextureNode GetTextureAt(int location);
-		void RemoveTextureAt(int location);
+		void RemoveTextureAt(size_t location);
 		void RemoveAllTextures();
 		int LoadFile(std::string location);
 		int TotalLoadedTextures();
